loop over versions, create class, create one obj per version
